% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMVDataset.R
\name{activate}
\alias{activate}
\title{Activate Dataset}
\usage{
activate(dataset)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}
}
\value{
MicroVis dataset (mvdata object)
}
\description{
Activate Dataset
}
