% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisAlphaDiv.R
\name{adiv}
\alias{adiv}
\title{Alpha Diversity Calculation}
\usage{
adiv(dataset = NULL, method = "chao1", rooted = F)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset}

\item{method}{Method for alpha diversity calculation. One of either "chao1",
"shannon", "simpson", "invsimpson", or "pd". Defaults to "chao1"}

\item{rooted}{(Optional) If method is set to "pd", calculate rooted or unrooted
Faith's phylogenetic diversity index? Defaults to FALSE}
}
\value{
Table with the alpha diversity index for each sample
}
\description{
Alpha Diversity Calculation
}
