% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASVtoTaxa.R
\name{agglomTaxa}
\alias{agglomTaxa}
\title{Agglomerate taxa into parent taxa}
\usage{
agglomTaxa(taxa_data, abundance_table, from_rank, to_rank)
}
\arguments{
\item{taxa_data}{Taxonomy data with abundance tables and taxa names table}

\item{abundance_table}{Abundance table}

\item{from_rank}{Starting rank (must be lower than "to_rank")}

\item{to_rank}{Destination rank (must be higher than "from_rank")}
}
\value{
List of parent taxa at the "to_rank" rank for each taxon in "abundance_table"
}
\description{
Agglomerate taxa into parent taxa
}
