% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASVtoTaxa.R
\name{ASVtoTaxa}
\alias{ASVtoTaxa}
\title{Translate ASV Numbers to Taxa Names}
\usage{
ASVtoTaxa(taxa_data, asvlist, taxa_rank = NULL, uniqueOnly = F)
}
\arguments{
\item{taxa_data}{Taxonomy data (list object) from a dataset containing
asv-taxonomy reference table.}

\item{asvlist}{List of ASV numbers to look up in taxa_data.}

\item{taxa_rank}{(Optional) The desired rank at which to look up ASV numbers.
Will default to the active rank if not specified.}

\item{uniqueOnly}{(Optional) If multiple taxa map to the same parent taxon,
only return the name of that parent taxon once. Default is FALSE}
}
\value{
List of taxa corresponding to the ASVs in the same order.
}
\description{
Translate ASV Numbers to Taxa Names
}
