% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settingSwitches.R
\name{autonametoggle}
\alias{autonametoggle}
\title{Automatically Name Results}
\usage{
autonametoggle()
}
\value{
Does not return anything
}
\description{
Cycles between naming results automatically or prompting user
    for a name
}
