% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisBetaDiv.R
\name{bdiv}
\alias{bdiv}
\title{Beta Diversity Analysis}
\usage{
bdiv(
  dataset = NULL,
  factor = NULL,
  stratifier = NULL,
  method = "bray",
  weighted = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset}

\item{factor}{Factor along which to conduct beta-diversity analysis}

\item{stratifier}{(Optional) A second factor along which to stratify groups}

\item{method}{Method for dissimilarity calculation. One of either "bray",
"euclidean", "jaccard", "unifrac", "manhattan", "canberra", "clark",
"kulczynski", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao", "mahalanobis", "chisq" or "chord"}

\item{weighted}{(Optional) If method is set to "unifrac", whether to perform
weighted or unweighted unifrac. Defaults to FALSE}
}
\value{
List of data tables containing various beta-diversity results
}
\description{
Beta Diversity Analysis
}
