% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisUnivar.R
\name{calcUniVar}
\alias{calcUniVar}
\title{Core Univariate Analysis Function}
\usage{
calcUniVar(
  data,
  factor,
  stratifiers = NULL,
  features,
  param = F,
  pairwise_comparisons = NULL
)
}
\arguments{
\item{data}{Data table containing metadata and count values with samples as
rows and metadata factors or features as columns}

\item{factor}{Factor along which to form groups from the samples and perform
the statistical analysis}

\item{stratifiers}{(Optional) One or two factors to stratify the groups by}

\item{features}{List of features to analyze. Defaults to all}

\item{param}{Run parametrized or nonparametrized univariate analysis? Default
is FALSE}

\item{pairwise_comparisons}{For 3 or more groups, list of vectors of pairwise
comparisons to conduct. Defaults to all possible pairwise comparisons}
}
\value{
List containing a table for overall statistics results and the features
    that were skipped due to having the same mean. For analyses with 3 or more
    groups, also contains a second table with the pairwise statistics results
}
\description{
Core Univariate Analysis Function
}
