% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize a Single-Word String}
\usage{
capitalize(string)
}
\arguments{
\item{string}{Single-word string.}
}
\value{
String with first letter capitalized and all other letters in lowercase.
}
\description{
Capitalize a Single-Word String
}
