% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeFactor.R
\name{changeFactor}
\alias{changeFactor}
\title{Switch to a Different Factor}
\usage{
changeFactor(dataset = NULL, factor_name = NULL, temp = F)
}
\arguments{
\item{dataset}{Microvis dataset (mvdata).}

\item{factor_name}{(Optional) Name of factor to switch to.}

\item{temp}{If set to TRUE, this will not update the active dataset.}
}
\value{
Dataset with new primary factor
}
\description{
Switch to a Different Factor
}
