% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsState.R
\name{changerank}
\alias{changerank}
\title{Change the active rank of a dataset}
\usage{
changerank(dataset = NULL, rank = NULL)
}
\arguments{
\item{dataset}{Microvis dataset. Default is the active dataset}

\item{rank}{(Optional) Specify rank to change to. If none is specified, the
function will ask}
}
\value{
MicroVis dataset with the active rank changed to the selected rank
}
\description{
Change the active rank of a dataset
}
