% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDataset.R
\name{checkColors}
\alias{checkColors}
\title{Check Colors of Groups in a Dataset}
\usage{
checkColors(dataset = NULL, assigned = FALSE)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}

\item{assigned}{(Optional) Whether or not to show what groups the colors
are assigned to. Default is FALSE.}
}
\value{
Plot of colors with/without the associated groups.
}
\description{
Check Colors of Groups in a Dataset
}
