% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisLMEM.R
\name{checkFormula}
\alias{checkFormula}
\title{Check that a formula is valid for a dataset}
\usage{
checkFormula(dataset, formula)
}
\arguments{
\item{dataset}{MicroVis dataset.}

\item{formula}{Formula for linear model.}
}
\value{
Formula if it is valid, or NULL if formula is not valid.
}
\description{
Check that a formula is valid for a dataset
}
