% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkGroups.R
\name{checkGroups}
\alias{checkGroups}
\title{Check that each group has at least a certain number of samples}
\usage{
checkGroups(
  dataset = NULL,
  factor = NULL,
  stratifiers = NULL,
  min_n = 3,
  verbose = T
)
}
\arguments{
\item{dataset}{MicroVis dataset. Default is the active dataset}

\item{factor}{Factor along which to group the samples}

\item{stratifiers}{Stratifiers along which to stratify the samples}

\item{min_n}{Minumum number of samples expected in each group}

\item{verbose}{If set to TRUE, prints out processing text}
}
\value{
TRUE or FALSE indicating whether any groups in the specified factor
   have fewer than the specified number of samples
}
\description{
Check that each group has at least a certain number of samples
}
