% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseFactors.R
\name{chooseFactors}
\alias{chooseFactors}
\title{Choose Factors for the Dataset}
\usage{
chooseFactors(dataset)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}
}
\value{
Microvis dataset with factors data
}
\description{
Choose Factors for the Dataset
}
