% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFiltering.R
\name{chooseGrps}
\alias{chooseGrps}
\title{Choose Groups to be Analyzed in Each Factor}
\usage{
chooseGrps(dataset = NULL, factor_names = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{factor_names}{(Optional) Names of the factors for which to specify groups}
}
\value{
MicroVis dataset (mvdata object) with updated "subset" lists for each
    of its factors, and an updated list of samples for runSampleFilter() to
    remove based on the groups chosen for analysis
}
\description{
Choose Groups to be Analyzed in Each Factor
}
