% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFiltering.R
\name{chooseSamples}
\alias{chooseSamples}
\title{Select specific samples to analyze}
\usage{
chooseSamples(dataset = NULL, samples, includeLowQual = F)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{samples}{List of samples to analyze}

\item{includeLowQual}{Include samples even if they are low quality? Defaults
to FALSE}
}
\value{
Dataset with updated list of samples to ignore
}
\description{
Select specific samples to analyze
}
