% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeASVtab.R
\name{cleanASVs}
\alias{cleanASVs}
\title{Clean ASVs in ASV Dataframe}
\usage{
cleanASVs(taxa_names_tab)
}
\arguments{
\item{taxa_names_tab}{Processed ASV dataframe}
}
\value{
Returns an ASV dataframe with taxa names that make more sense and are
    easier to deal with.
}
\description{
Rename NA taxa to refer to their highest resolution of identification
    and make sure species names are prepended with the genus name
}
