% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanData}
\alias{cleanData}
\title{Clean up a metadata factor column in a merged data table}
\usage{
cleanData(merged_data, factor)
}
\arguments{
\item{merged_data}{Datatframe of merged metadata and abundance table}

\item{factor}{Factor to "clean" by ensuring its column is a factor with only
levels that the filtered dataset contains}
}
\value{
Merged metadata+abundance table with the metadata column of the
    specified factor re-leveled to only include levels that exist in the
    column
}
\description{
Clean up a metadata factor column in a merged data table
}
