% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanGroups}
\alias{cleanGroups}
\title{Clean Metadata Factors}
\usage{
cleanGroups(melted_data, factor_name, verbose = T)
}
\arguments{
\item{melted_data}{Dataframe with metadata and abundance data.}

\item{factor_name}{Factor to refactor in metadata.}

\item{verbose}{If TRUE (default), print out the count of samples in each group
of the factor and warn the user if fewer than 3 samples are in a group.}
}
\value{
A nested list containing the sample count for each group in the factor.
}
\description{
Clean Metadata Factors
}
