% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxonomy.R
\name{cleanUnkTaxa}
\alias{cleanUnkTaxa}
\title{Rename Taxa with NA and at Species Level}
\usage{
cleanUnkTaxa(dataset = NULL)
}
\arguments{
\item{dataset}{Unprocessed MicroVis dataset (mvdata object) that still contains
the unranked abundance table.}
}
\value{
MicroVis dataset (mvdata object) with cleaned taxonomy names in its
    asv-taxonomy reference table
}
\description{
Rename Taxa with NA and at Species Level
}
