% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorGroups.R
\name{colorGrps}
\alias{colorGrps}
\title{Color Groups}
\usage{
colorGrps(dataset, new_order.unlisted = NULL, clrd_grps = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{new_order.unlisted}{Single list of any new groups across all factors
with their ordering retained}

\item{clrd_grps}{Named list of already colored groups and their colors}
}
\value{
MicroVis dataset (mvdata object) with a colors attribute containing
    the color assignment for each group
}
\description{
Color Groups
}
