% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDataset.R
\name{countFeatures}
\alias{countFeatures}
\title{Get Number of Included Samples in a Dataset}
\usage{
countFeatures(dataset = NULL)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}
}
\value{
Number of features in a dataset at its active rank.
}
\description{
Get Number of Included Samples in a Dataset
}
