% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDataset.R
\name{countSamples.base}
\alias{countSamples.base}
\title{Core function for calculating sample sizes}
\usage{
countSamples.base(
  metadata,
  factors,
  stratifiers = NULL,
  min_n = 3,
  dataset = NULL,
  verbose = T
)
}
\arguments{
\item{metadata}{Metadata of the dataset}

\item{factors}{Factors along which samples will be counted in groups}

\item{stratifiers}{Factors along which groups will be stratified}

\item{min_n}{Threshold at which to warn user of a small sample size in a group
or a stratified group.}

\item{dataset}{(Optional) Dataset is passed to relay information about factors that is
otherwise lost.}

\item{verbose}{Whether or not to print out the size report for the groups/
stratified groups.}
}
\value{
Returns a nested list of the sample size for each group/stratified group
}
\description{
Core function for calculating sample sizes
}
