% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDataset.R
\name{countSamples}
\alias{countSamples}
\title{Get Number of Included Samples in a Dataset}
\usage{
countSamples(
  dataset = NULL,
  factors = NULL,
  stratifiers = NULL,
  getSizes = F,
  min_n = 3,
  verbose = T
)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}

\item{factors}{Factors along which samples will be split into groups.
Default is the active factor of the dataset.}

\item{stratifiers}{(Optional) Factors along which groups will be stratified.
Default is none.}

\item{getSizes}{Whether or not to return a nested list with the sizes of each
groups. Default is FALSE.}

\item{min_n}{Passed along to countSamples.base (non-exported function)}

\item{verbose}{(Optional) If TRUE, will print out sizes of each group in each
factor.}
}
\value{
If "getSizes" flag is TRUE, will report group sizes. If not, returns
    nothing.
}
\description{
Get Number of Included Samples in a Dataset
}
