% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeFeatures.R
\name{excludeFeatures}
\alias{excludeFeatures}
\title{Exclude specific features at a specified taxonomic rank. Only for taxonomic
    datasets with multiple taxonomic ranks.}
\usage{
excludeFeatures(dataset = NULL, features, rank = NULL, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{features}{Vector of features (at any rank) to select}

\item{rank}{(Optional) Specify a rank from which to select features. Searches
through all ranks by default}

\item{temp}{This parameter has no use in this function and can be removed}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
Dataset with list of excluded features that is passed to runFeatureFilter
}
\description{
Exclude specific features at a specified taxonomic rank. Only for taxonomic
    datasets with multiple taxonomic ranks.
}
