% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{filterLowRelAbun}
\alias{filterLowRelAbun}
\title{Filter Low Relative Abundance Features}
\usage{
filterLowRelAbun(dataset = NULL, top = NULL, min_relabun = NULL, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{top}{(Optional) Top n number of features to be kept based relative
abundance. Specifying a value for this will override the min_relabun
value.}

\item{min_relabun}{(Optional) (Default) Relative abundance threshold. Features
with prevalence lower than this threshold will be filtered out.}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
MicroVis dataset (mvdata object) with updated low relative abundance
    filtering parameters.
}
\description{
Filter Low Relative Abundance Features
}
