% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{filterLowTotAbun}
\alias{filterLowTotAbun}
\title{Filter Low Total Abundance Features}
\usage{
filterLowTotAbun(dataset = NULL, top = NULL, min_totabun = NULL, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{top}{(Optional) Top n number of features to be kept based on total abundance.
Specifying a value for this will override the min_totabun value.}

\item{min_totabun}{(Optional) (Default) Total abundance threshold. Features
with total abundance lower than this threshold will be filtered out.}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
MicroVis dataset (mvdata object) with updated low total abundance
    filtering parameters.
}
\description{
Filter Low Total Abundance Features
}
