% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{filterLowVar}
\alias{filterLowVar}
\title{Filter Low Variance Features}
\usage{
filterLowVar(
  dataset = NULL,
  top = NULL,
  var_type = "sd",
  low_var_percentile = NULL,
  silent = F
)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{top}{(Optional) Top n number of features to be kept based on relative
abundance. Specifying a value for this will override the min_relabun value.}

\item{var_type}{Method for assessing variance of samples. Current choices are
standard deviation ('sd') and interquartile range ('iqr')}

\item{low_var_percentile}{(Optional) (Default) Variance threshold. Features
with variance (standard deviation) lower than this threshold will be
filtered out.}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
MicroVis dataset (mvdata object) with updated low variance filtering
    parameters.
}
\description{
Filter Low Variance Features
}
