% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{findSigFisher}
\alias{findSigFisher}
\title{Run Fisher Test on Features}
\usage{
findSigFisher(dataset, fts, lowabun_thresh = 0, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{fts}{List of features to test.}

\item{lowabun_thresh}{Abundance value above which a feature is considered to
be present in a given sample.}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
List of features that had a statistically significant Fisher test
    when performed across the subsetted groups of the active factor.
}
\description{
Run Fisher test on features that are labeled for filtering. This
    test is run on presence-absence data and not abundance data.
}
