% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFoldChange.R
\name{foldChange}
\alias{foldChange}
\title{Calculate Fold Change}
\usage{
foldChange(
  dataset = NULL,
  factor = NULL,
  ref_group = NULL,
  comparison_groups = NULL,
  ranks = NULL,
  allRanks = F,
  pairSamples = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset}

\item{factor}{Factor along which to conduct beta-diversity analysis}

\item{ref_group}{Reference group for fold change computation. This will be the
denominator. User will be asked if none is specified}

\item{comparison_groups}{Comparison group for fold change computation. This
will be the numerator. User will be asked if none is specified}

\item{ranks}{Ranks at which to calculate fold change of features}

\item{allRanks}{If set to TRUE, will compute feature fold change at all ranks.
Defaults to FALSE}

\item{pairSamples}{Should fold change be calculated on paired samples instead
of between the means of groups? Defaults to FALSE. This option is useful
when you have 2 or more time points from the same patient or something
similar}
}
\value{
Table with the fold change and log2(fold change) for the features
}
\description{
Calculate Fold Change
}
