% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFeatureCorr.R
\name{ftcor}
\alias{ftcor}
\title{Compute feature correlation matrices for each group and all groups together}
\usage{
ftcor(
  dataset1 = NULL,
  dataset2 = NULL,
  method = c("pearson", "kendall", "spearman"),
  factor = NULL,
  features1 = NULL,
  features2 = NULL
)
}
\arguments{
\item{dataset1}{MicroVis dataset. Defaults to the active dataset}

\item{dataset2}{(Optional) MicroVis dataset with features to correlate with
the dataset1}

\item{method}{Correlation method. One of either "pearson", "spearman", or
"kendall". Defaults to "spearman"}

\item{factor}{Factor to group samples by}

\item{features1}{(Optional) Specific features to consider from first dataset.
If only one dataset is being used, then these features will be correlated
with features of the same dataset}

\item{features2}{(Optional) Specific features to consider from second dataset.}
}
\value{
List of correlation matrices for each group with additional matrices
    containing p-values and adjusted p-values
}
\description{
Compute feature correlation matrices for each group and all groups together
}
