% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftRatios.R
\name{ftRatio}
\alias{ftRatio}
\title{Calculate ratio between two features}
\usage{
ftRatio(data, feature1, feature2)
}
\arguments{
\item{data}{Abundance table}

\item{feature1}{Numerator feature}

\item{feature2}{Denominator feature}
}
\value{
Ratio of the two specified features
}
\description{
Calculate ratio between two features
}
