% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftRatios.R
\name{ftRatiostab}
\alias{ftRatiostab}
\title{Generate a table of feature ratios}
\usage{
ftRatiostab(
  melted_data,
  feature1list,
  feature2list,
  factor = NA,
  aggregateBefore = F
)
}
\arguments{
\item{melted_data}{Melted metadata and abundance table}

\item{feature1list}{List of numerator features}

\item{feature2list}{List of denominator features}

\item{factor}{(Optional) Factor along which to group samples to calculate an
aggregated ratio}

\item{aggregateBefore}{Aggregate samples before calculating ratios? Stats will
not be calculated between groups in this case.}
}
\value{
Table of feature ratios between specified features
}
\description{
Generate a table of feature ratios
}
