% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDataset.R
\name{getdata}
\alias{getdata}
\title{Get Data with/without Attached Metadata}
\usage{
getdata(dataset = NULL, rank = NULL, group = NULL, metadata = T)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}

\item{rank}{(Optional) Rank to get abundance values of. Default is active rank.}

\item{group}{This isn't used and can be removed}

\item{metadata}{(Optional) Whether or not to attach metadata.}
}
\value{
Abundance data at specified rank with/without metadata.
}
\description{
Get Data with/without Attached Metadata
}
