% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getDiscriminatingFeatures}
\alias{getDiscriminatingFeatures}
\title{Get features that uniquely identify one group from other groups}
\usage{
getDiscriminatingFeatures(
  pairwise_statistics,
  unique_group,
  comparison_groups = NULL,
  alpha = 0.05
)
}
\arguments{
\item{pairwise_statistics}{Data frame of pairwise statistics for a comparison
of 3 or more groups}

\item{unique_group}{The group of interest -- features that are uniquely
differentially expressed in this group compared to the comparison groups
will be identified}

\item{comparison_groups}{The groups to compare the group of interest to.
Defaults to all other groups in the subsetted dataset}

\item{alpha}{Significance threshold. Defaults to 0.05}
}
\value{
List of features uniquely differentially expressed in "unique_group"
}
\description{
Get features that uniquely identify one group from other groups
}
