% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getFeatures}
\alias{getFeatures}
\title{Get list of features in a dataset}
\usage{
getFeatures(dataset = NULL, ranks = NULL, allRanks = F)
}
\arguments{
\item{dataset}{Dataset of interest. Default is the currently active dataset}

\item{ranks}{Vector of ranks at which to get the list of features}

\item{allRanks}{If set to TRUE, will get the list of all features at all ranks}
}
\value{
List of features in a dataset at specified ranks
}
\description{
Get list of features in a dataset
}
