% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisLMEM.R
\name{getFormulaVars}
\alias{getFormulaVars}
\title{Extract Variables from a Formula}
\usage{
getFormulaVars(formula)
}
\arguments{
\item{formula}{Formula as a string.}
}
\value{
Variables extracted from the formula in a vector
}
\description{
Extract Variables from a Formula
}
