% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLowestRank}
\alias{getLowestRank}
\title{Get Lowest Rank in a Dataset}
\usage{
getLowestRank(dataset = NULL)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}
}
\value{
Lowest rank of the data in a dataset.
}
\description{
Get Lowest Rank in a Dataset
}
