% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisNetwork.R
\name{getNetwork}
\alias{getNetwork}
\title{Get the co-occurence network for a dataset}
\usage{
getNetwork(
  dataset = NULL,
  factor = NULL,
  rank = NULL,
  fts = NULL,
  method = c("se.mb", "se.glasso"),
  format = c("none", "network", "igraph")
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{(Optional) Factor to split data into groups}

\item{rank}{Rank of features to analyze. Defaults to the active rank}

\item{fts}{(Optional) Vector of feature names to analyze. Defaults to all the
features in the filtered dataset}

\item{method}{Co-occurence analysis method}

\item{format}{(Optional) Format of the network. One of "none", "igraph", or
"network"}
}
\value{
Co-occurence network for a dataset
}
\description{
Get the co-occurence network for a dataset
}
