% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASVtoTaxa.R
\name{getParentTaxa}
\alias{getParentTaxa}
\title{Get parent taxa}
\usage{
getParentTaxa(taxalist, from, to, taxanames)
}
\arguments{
\item{taxalist}{Vector of current taxa}

\item{from}{Rank of current taxa}

\item{to}{Rank of desired parent taxa}

\item{taxanames}{Table of all taxa names}
}
\value{
Vector of parent taxa for each taxa in "taxalist"
}
\description{
Get parent taxa
}
