% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRanks}
\alias{getRanks}
\title{Get All Ranks of a Dataset}
\usage{
getRanks(dataset = NULL)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}
}
\value{
List of ranks in a dataset.
}
\description{
Get All Ranks of a Dataset
}
