% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{glog}
\alias{glog}
\title{Generalized Log}
\usage{
glog(x, base = 10)
}
\arguments{
\item{x}{Any real number}

\item{base}{Log base. Defaults to 10}
}
\value{
Generalized log value of the input
}
\description{
Generalized Log
}
