% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settingSwitches.R
\name{keepSigFisher}
\alias{keepSigFisher}
\title{Function for setting whether or not to run Fisher tests to prune filtering}
\usage{
keepSigFisher(keep = T)
}
\arguments{
\item{keep}{Keep features with significant Fisher test if they would have been
filtered out otherwise? Default is TRUE}
}
\description{
Function for setting whether or not to run Fisher tests to prune filtering
}
