% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listsigs}
\alias{listsigs}
\title{List Significant Features of a Dataset in Current State}
\usage{
listsigs(
  dataset = NULL,
  factor = NULL,
  ranks = NULL,
  allRanks = F,
  alpha = 0.05,
  silent = F,
  dataset_name = NULL
)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}

\item{factor}{(Optional) Factor along which to perform statistical analysis.
Default is the active factor of the dataset}

\item{ranks}{(Optional) Ranks at which to get the significant features. Default
is the active rank of the dataset}

\item{allRanks}{If set to TRUE, will get significant features at all ranks
of the dataset}

\item{alpha}{(Optional) Significance threshold. Default is 0.05.}

\item{silent}{If set to TRUE, output text will not be printed}

\item{dataset_name}{(Optional) Name of the dataset being passed so that results
can be saved to that dataset}
}
\value{
List of significant features.
}
\description{
List Significant Features of a Dataset in Current State
}
