% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{listUniques}
\alias{listUniques}
\title{List features uniquely over/under-expressed in a specified group}
\usage{
listUniques(
  dataset = NULL,
  factor = NULL,
  groups = NULL,
  rank = NULL,
  alpha = 0.05,
  dataset_name = NULL
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{Factor along which to perform univariate analysis. Defaults to
the active factor}

\item{groups}{Group(s) to identify unique features in. Cycles through all
groups by default}

\item{rank}{Rank at which to analyze features}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{dataset_name}{(Not recommended) Name of the dataset to save statistics
to. This should not need to be used by users since the function can
determine the name of the dataset directly passed to it, but not when
it is called within another function.}
}
\value{
A grouped list of features unique to each group of a factor in a dataset
}
\description{
List features uniquely over/under-expressed in a specified group
}
