% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctionalData.R
\name{loadFxnlFile}
\alias{loadFxnlFile}
\title{Load pathways file}
\usage{
loadFxnlFile(path_to_fxnl = NA, metadata = NULL)
}
\arguments{
\item{path_to_fxnl}{(Optional) full path to pathway csv file}

\item{metadata}{Metadata dataframe}
}
\value{
Loaded pathway abundance dataframe
}
\description{
Load pathways file
}
