% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadMetadata.R
\name{loadMDFile}
\alias{loadMDFile}
\title{Load metadata file}
\usage{
loadMDFile(path_to_metadata = NA)
}
\arguments{
\item{path_to_metadata}{(Optional) full path to metadata csv file}
}
\value{
Loaded metadata dataframe
}
\description{
Load metadata file
}
