% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTaxonomy.R
\name{loadTaxaFile}
\alias{loadTaxaFile}
\title{Load Taxonomy File}
\usage{
loadTaxaFile(path_to_taxa = NA, metadata = NULL, combineDupes = T)
}
\arguments{
\item{path_to_taxa}{(Optional) full path to taxonomy csv file}

\item{metadata}{Metadata dataframe}

\item{combineDupes}{If set to TRUE (default), MicroVis will try to combine
duplicate features.}
}
\value{
Loaded taxonomy data - a list of:
    1) orig: A sample-by-ASV (rows-by-columns) original abundance dataframe
    that will not be processed.
    2) A temporary copy of orig that will be used in downstream processing
    3) An ASV-by-rank dataframe with ASVs and their assigned taxa at each
    taxonomic rank
}
\description{
Load Taxonomy File
}
