% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetTranslations.R
\name{makeAldex}
\alias{makeAldex}
\title{Make an ALDEx2 CLR Object}
\usage{
makeAldex(dataset = NULL, factor = NULL, rank = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Default is the active dataset}

\item{factor}{Factor to use as condition in making the ALDEx2 object}

\item{rank}{Rank at which to create ALDEx2 object}
}
\value{
An aldex.clr object
}
\description{
Make an ALDEx2 CLR Object
}
