% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeASVtab.R
\name{makeASVtab}
\alias{makeASVtab}
\title{Make ASV Table}
\usage{
makeASVtab(taxa_names_tab)
}
\arguments{
\item{taxa_names_tab}{List of unprocessed taxa that may or may not already
be split into columns for each rank}
}
\value{
Dataframe with taxa in rows, split into columns for each taxonomic rank.
}
\description{
Generates an ASV table based on the taxa present in the taxonomy
    csv file. Will attempt to guess how many ranks are represented in the
    taxa names if they are not explicitly stated in the file.
}
