% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetTranslations.R
\name{makeDeseq}
\alias{makeDeseq}
\title{Make a DESeq object from a MicroVis dataset}
\usage{
makeDeseq(dataset = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Default is the active dataset}
}
\value{
DESeq object
}
\description{
Make a DESeq object from a MicroVis dataset
}
