% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetTranslations.R
\name{makePS}
\alias{makePS}
\title{Make a phyloseq object from a MicroVis dataset}
\usage{
makePS(dataset = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Default is the active dataset}
}
\value{
Phyloseq object
}
\description{
Make a phyloseq object from a MicroVis dataset
}
