% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRankTables.R
\name{makeRankTabs}
\alias{makeRankTabs}
\title{Make Abundance Tables at Each Rank}
\usage{
makeRankTabs(ft_data)
}
\arguments{
\item{ft_data}{Feature data containing table of feature names - if taxa,
then separated by rank - an original abundance table, and an unranked
abundance table. If features are not taxa, then this function returns
the input unchanged.}
}
\value{
Feature data with table of feature names, an original abundance table
    and processed abundance tables at each taxonomy rank.
}
\description{
Make Abundance Tables at Each Rank
}
