% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetTranslations.R
\name{makeTaxMap}
\alias{makeTaxMap}
\title{Make a taxmap object from a MicroVis dataset}
\usage{
makeTaxMap(dataset = NULL, unfiltered = F, ftlist = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset}

\item{unfiltered}{Use the unfiltered MicroVis dataset?}

\item{ftlist}{List of features to include into the taxmap object}
}
\value{
taxmap object
}
\description{
Make a taxmap object from a MicroVis dataset
}
