% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultSaving.R
\name{multisave}
\alias{multisave}
\title{Determine how to save multiple figures}
\usage{
multisave(save_one_all = NULL)
}
\arguments{
\item{save_one_all}{Variable to determine how to save figures}
}
\value{
save_one_all
}
\description{
Determine how to save multiple figures
}
