% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisAldex.R
\name{mvaldex}
\alias{mvaldex}
\title{ALDEx2 Differential Abundance Analysis}
\usage{
mvaldex(dataset = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Default is the active dataset}
}
\value{
Results of ALDEx2 differential abundance analysis
}
\description{
ALDEx2 Differential Abundance Analysis
}
