% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeSeq_Ziln.R
\name{mvcalcNormFactors}
\alias{mvcalcNormFactors}
\title{Cumulative sum scaling (css) normalization factors}
\usage{
mvcalcNormFactors(x, p = mvcumNormStatFast(x))
}
\arguments{
\item{x}{Abundance table with samples as columns and features as rows}

\item{p}{The pth quantile.}
}
\value{
Vector of the sum up to and including a sample's pth quantile.
}
\description{
Return a vector of the the sum up to and including a quantile.
}
\examples{

data(mouseData)
head(calcNormFactors(mouseData))

}
