% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeSeq_Ziln.R
\name{mvcalcPosComponent}
\alias{mvcalcPosComponent}
\title{Positive component}
\usage{
mvcalcPosComponent(mat, mod, weights)
}
\arguments{
\item{mat}{A matrix of normalized counts}

\item{mod}{A model matrix}

\item{weights}{Weight matrix for samples and counts}
}
\description{
Fit the positive (log-normal) component
}
