% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeSeq_Ziln.R
\name{mvcalcShrinkParameters}
\alias{mvcalcShrinkParameters}
\title{Calculate shrinkage parameters}
\usage{
mvcalcShrinkParameters(fit, coef, mins2, exclude = NULL)
}
\arguments{
\item{fit}{A matrix of fits as outputted by calcZeroComponent or calcPosComponent}

\item{coef}{Coefficient of interest}

\item{mins2}{minimum variance estimate}

\item{exclude}{Vector of features to exclude when shrinking}
}
\description{
Calculate the shrunken variances and variance of parameters of interest across features.
}
