% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeSeq_Ziln.R
\name{mvcalcZeroComponent}
\alias{mvcalcZeroComponent}
\title{Zero component}
\usage{
mvcalcZeroComponent(mat, mod, weights)
}
\arguments{
\item{mat}{A matrix of normalized counts}

\item{mod}{A model matrix}

\item{weights}{Weight matrix for samples and counts}
}
\description{
Fit the zero (logisitic) component
}
