% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineDatasets.R
\name{mvcombine}
\alias{mvcombine}
\title{Merge Two Datasets with Different Sample Sets}
\usage{
mvcombine(path_to_folder = NA)
}
\arguments{
\item{path_to_folder}{Folder with project directory.}
}
\value{
Merged dataset with samples from both datasets kept separate and all
    the features from both datasets.
}
\description{
Merge Two Datasets with Different Sample Sets
}
