% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisPermanova.R
\name{mvdist}
\alias{mvdist}
\title{Distance Calculation}
\usage{
mvdist(
  dataset = NULL,
  method = "bray",
  weighted = F,
  allFactors = T,
  factors = NULL
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset}

\item{method}{Method for dissimilarity calculation. One of either "bray",
"euclidean", "jaccard", "unifrac", "manhattan", "canberra", "clark",
"kulczynski", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao", "mahalanobis", "chisq" or "chord"}

\item{weighted}{(Optional) If method is set to "unifrac", whether to perform
weighted or unweighted unifrac. Defaults to FALSE}

\item{allFactors}{When performing distance calculation, only include samples
in the active subgroups of all factors? Defaults to TRUE. Otherwise, user
can specify which factors using the factor parameter (explained below)}

\item{factors}{(Optional) Specify factors to filter samples by}
}
\value{
Dissimilarity matrix and other associated details of the calculation
}
\description{
Distance Calculation
}
