% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvhelp.R
\name{mvhelp}
\alias{mvhelp}
\title{List available functions}
\usage{
mvhelp(pages = c("filtering", "normalizing", "analysis", "plotting"))
}
\arguments{
\item{pages}{(Optional) Specify what function options to see with "filtering",
"normalizing", "analysis", "plotting", or "all"}
}
\value{
Prints a list of available functions with short descriptions
}
\description{
List available functions
}
