% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisLMEM.R
\name{mvLMEM}
\alias{mvLMEM}
\title{Linear Mixed Effects Model Discriminant Analysis}
\usage{
mvLMEM(
  dataset = NULL,
  formula = NULL,
  exclude = NULL,
  rank = NULL,
  zero.handling = "pseudo-count",
  alpha = 0.05,
  is.winsor = T
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset.}

\item{formula}{Formula for linear model. Defaults to simple linear model of all covariates.}

\item{exclude}{Factors/covariates to exclude in linear model.}

\item{rank}{Rank at which to conduct analysis.}

\item{zero.handling}{(From linda function) A character string of 'pseudo-count' or 'imputation' indicating the zero handling method used when feature.dat is 'count'. If 'pseudo-count', apseudo.cnt will be added to each value in feature.dat. If 'imputation', then we use the imputation approach using the formula in the referenced paper. Basically, zeros are imputed with values proportional to the sequencing depth. When feature.dat is 'proportion', this parameter will be ignored and zeros will be imputed by half of the minimum for each feature.}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{is.winsor}{Whether to replace outliers (using winsorization)}
}
\value{
Results of linear modeling with log2fc, p-values, adjusted p-values for each covariate.
}
\description{
Linear Mixed Effects Model Discriminant Analysis
}
