% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvload.R
\name{mvload}
\alias{mvload}
\title{MicroVis Data Loading}
\usage{
mvload(
  path_to_folder = NULL,
  path_to_metadata = NA,
  path_to_taxa = NA,
  path_to_fxnl = NA,
  autoProcess = T,
  combineDupes = T,
  combineDataSets = F
)
}
\arguments{
\item{path_to_folder}{(Optional) full path to directory containing csv files}

\item{path_to_metadata}{(Optional) Path to metadata csv file}

\item{path_to_taxa}{(Optional) Path to taxonomy abundance csv file}

\item{path_to_fxnl}{(Optional) Path to functional abundance csv file}

\item{autoProcess}{If set to TRUE (default), MicroVis will automatically
process datasets using default normalization and filtering parameters.
Datasets can still be re-processed by the user afterwards.}

\item{combineDupes}{If set to TRUE (default), MicroVis will try to combine
duplicate features.}

\item{combineDataSets}{If set to TRUE, MicroVis will ask for a second set of
data and combine both datasets into one.}
}
\value{
List of successfully loaded datasets with/without processing.
}
\description{
Initial function for loading taxonomy and/or pathway data with
    associated metadata into MicroVis
}
