% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeDatasets.R
\name{mvmerge}
\alias{mvmerge}
\title{Merge Two Datasets with Overlapping Samples}
\usage{
mvmerge(
  dataset1,
  dataset2,
  dataset1_name = NULL,
  dataset2_name = NULL,
  features1 = NULL,
  features2 = NULL,
  keepFiltered = F,
  sigsOnly1 = F,
  sigsOnly2 = F,
  alpha = 0.05
)
}
\arguments{
\item{dataset1}{First MicroVis dataset}

\item{dataset2}{Second MicroVis dataset}

\item{dataset1_name}{(Optional) What to name the first dataset.}

\item{dataset2_name}{(Optional) What to name the second dataset.}

\item{features1}{(Optional) Features to choose from first dataset.}

\item{features2}{(Optional) Features to choose from second dataset.}

\item{keepFiltered}{Whether or not to keep the filtered-out taxa that are lumped
into the "Other" feature by runFeatureFilter().}

\item{sigsOnly1}{Whether or not to only consider the significant features of
the first dataset. Default is FALSE.}

\item{sigsOnly2}{Whether or not to only consider the significant features of
the second dataset. Default is FALSE.}

\item{alpha}{Significance threshold. Default is 0.05.}
}
\value{
Merged dataset with the samples that were shared across both datasets
    and all the features selected from both datasets.
}
\description{
Merge Two Datasets with Overlapping Samples
}
