% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisNetwork.R
\name{mvNetwork}
\alias{mvNetwork}
\title{Construct an association network between features}
\usage{
mvNetwork(dataset = NULL, factor = NULL, rank = NULL, fts = NULL, fill = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{(Optional) Factor to split data into groups}

\item{rank}{Rank of features to analyze. Defaults to the active rank}

\item{fts}{(Optional) Vector of feature names to analyze. Defaults to all the
features in the filtered dataset}

\item{fill}{(Optional) How to choose the fill color of the nodes. Defaults to
the highest rank for taxonomic data, or a single color for other data.}
}
\value{
MicroVis dataset with network coefficients attached
}
\description{
Construct an association network between features
}
