% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltData.R
\name{mvstratify}
\alias{mvstratify}
\title{Make a melted data table grouped by strata}
\usage{
mvstratify(
  dataset = NULL,
  factor = NULL,
  stratifiers = NULL,
  rank = NULL,
  min_n = 3,
  verbose = T
)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}

\item{factor}{Main factor whose groups will be in each grouped data table}

\item{stratifiers}{Stratifier(s) to group the melted data table by}

\item{rank}{Rank of dataset to choose features from}

\item{min_n}{Passed to cleanGroups()}

\item{verbose}{If set to TRUE, will print processing text}
}
\value{
Grouped, melted data table of metadata and abundance at a specified
    rank
}
\description{
Make a melted data table grouped by strata
}
