% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisBetaDiv.R
\name{mvunifrac}
\alias{mvunifrac}
\title{Unifrac Distance Calculation}
\usage{
mvunifrac(dataset = NULL, weighted = F, normalized = T)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset}

\item{weighted}{Whether to perform weighted or unweighted unifrac.
Defaults to FALSE}

\item{normalized}{Normalize dissimilarities to a range of 0 to 1? Defaults to
TRUE}
}
\value{
"dist" object of unifrac dissimilarity values between samples
}
\description{
Unifrac Distance Calculation
}
