% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisZimm.R
\name{mvZig}
\alias{mvZig}
\title{Zero-inflated Gaussian Fitting and Statistical Analysis of Differences}
\usage{
mvZig(dataset = NULL, factor = NULL, rank = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{Factor to analyze by. Defaults to the active factor}

\item{rank}{Rank to select features from. Defaults to active rank}
}
\value{
metagenomeSeq fitFeatureModel results
}
\description{
Zero-inflated Gaussian Fitting and Statistical Analysis of Differences
}
