% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metagenomeSeq_Zig.R
\name{mvzigControl}
\alias{mvzigControl}
\alias{settings2}
\title{Settings for the fitZig function}
\usage{
mvzigControl(
  tol = 1e-04,
  maxit = 10,
  verbose = TRUE,
  dfMethod = "modified",
  pvalMethod = "default"
)
}
\arguments{
\item{tol}{The tolerance for the difference in negative log likelihood estimates for a feature to remain active.}

\item{maxit}{The maximum number of iterations for the expectation-maximization algorithm.}

\item{verbose}{Whether to display iterative step summary statistics or not.}

\item{dfMethod}{Either 'default' or 'modified' (by responsibilities).}

\item{pvalMethod}{Either 'default' or 'bootstrap'.}
}
\value{
The value for the tolerance, maximum no. of iterations, and the verbose warning.
}
\description{
Settings for the fitZig function
}
\examples{
control =  zigControl(tol=1e-10,maxit=10,verbose=FALSE)

}
