% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisZimm.R
\name{mvZimm}
\alias{mvZimm}
\title{Zero-Inflated Mixed Model Analysis}
\usage{
mvZimm(dataset = NULL, factor = NULL, rank = NULL, dataset_name = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{Factor to analyze by. Defaults to the active factor}

\item{rank}{Rank to select features from. Defaults to active rank}

\item{dataset_name}{(Not Recommended) Name of the dataset to save results
to. This should not need to be used by users since the function can
determine the name of the dataset directly passed to it, but not when
it is called within another function.}
}
\value{
MicroVis dataset containing results of LEfSe analysis
}
\description{
Zero-Inflated Mixed Model Analysis
}
