% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultSaving.R
\name{nameAnalysis}
\alias{nameAnalysis}
\title{Automatically name the analysis based on factors and the active factor}
\usage{
nameAnalysis(factors, active_factor, facets = NULL)
}
\arguments{
\item{factors}{List of factors of the dataset}

\item{active_factor}{The name of the active factor of the dataset}

\item{facets}{(Optional) Any facets used in an analysis}
}
\value{
Name of the analysis as a string
}
\description{
Automatically name the analysis based on factors and the active factor
}
