% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nameStratification}
\alias{nameStratification}
\title{Standardized naming of the stratification of an analysis}
\usage{
nameStratification(stratifiers)
}
\arguments{
\item{stratifiers}{Factors by which an analysis is stratified}
}
\value{
Stratified name as a string
}
\description{
Standardized naming of the stratification of an analysis
}
