% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisNetwork.R
\name{networkClusters}
\alias{networkClusters}
\title{Get feature clusters determined by network analysis}
\usage{
networkClusters(
  dataset = NULL,
  factor = NULL,
  rank = NULL,
  fts = NULL,
  method = c("se.mb", "se.glasso")
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{(Optional) Factor to split data into groups}

\item{rank}{Rank of features to analyze. Defaults to the active rank}

\item{fts}{(Optional) Vector of feature names to analyze. Defaults to all the
features in the filtered dataset}

\item{method}{Co-occurence analysis method}
}
\value{
List of feature clusters
}
\description{
Get feature clusters determined by network analysis
}
