% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalizeTable}
\alias{normalizeTable}
\title{Normalize a Standalone Abundance Table}
\usage{
normalizeTable(
  abundance_table,
  sample_scale = NULL,
  feature_scale = NULL,
  transformation = NULL,
  sum_scale = 1e+06,
  log_base = 10,
  silent = F
)
}
\arguments{
\item{abundance_table}{Abundance table with samples as rows and features as
columns}

\item{sample_scale}{Sample scaling method to be used. Choices are "sum", "none",
"relative", "median", "range", "centered", and "standardized". Default is
"sum".}

\item{feature_scale}{Feature scaling method to be used. Choices are "sum", "none",
"relative", "median", "range", "centered", and "standardized". Default is
"sum".}

\item{transformation}{Transformation method to be used. Choices are "glog"
(generalized log), "none", "pseudolog", and "log".}

\item{sum_scale}{Number to scale samples/features by. Default is 1 million.}

\item{log_base}{Base for log transformations. Default is 10}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
Normalized abundance table.
}
\description{
Normalize a Standalone Abundance Table
}
