% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderGroups.R
\name{orderGroups}
\alias{orderGroups}
\title{Order Groups}
\usage{
orderGroups(dataset, reorder = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{reorder}{Set this to true if you want to reorder groups after they have
already been ordered in initial loading (during mvload)}
}
\value{
MicroVis dataset with groups in each factor ordered according to user's
    input
}
\description{
Order Groups
}
