% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisPairedCor.R
\name{pairedCor}
\alias{pairedCor}
\title{Correlate abundance of features between paired samples}
\usage{
pairedCor(
  dataset = NULL,
  ids,
  compare,
  fts = NULL,
  rank = NULL,
  rthresh = 0.7,
  alpha = 0.05,
  padj = T,
  data.paired = NULL,
  groups = NULL
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{ids}{Column(s) that uniquely identify each pair of samples}

\item{compare}{Column with the 2 groups to correlate between. Each pair of
samples has a sample in each of the two groups in this column. If there
are more than 2 groups in the column, user will be asked to select 2 to
compare}

\item{fts}{Vector of features to correlate between the two sets}

\item{rank}{Taxonomic rank at which to perform the paired correlation. Defaults
to the second highest rank in the dataset}

\item{rthresh}{R value cutoff for coloring the figures of each figure.
Defaults to 0.7}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{padj}{Adjust the p-values for multiple testing? Defaults to TRUE}

\item{data.paired}{Provide a dataframe with data already paired according to
"ids" and "compare"}

\item{groups}{If providing a data.paired dataframe, must provide groups as well}
}
\value{
Scatter plot with trendline
}
\description{
Correlate abundance of features between paired samples
}
