% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisLMEM.R
\name{parseFormula}
\alias{parseFormula}
\title{Extract Terms from a Formula}
\usage{
parseFormula(formula)
}
\arguments{
\item{formula}{Formula as a string.}
}
\value{
Variables extracted from the formula in a vector
}
\description{
Extract Terms from a Formula
}
