% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseStratifiers.R
\name{parseStratifiers}
\alias{parseStratifiers}
\title{Parse Stratifiers}
\usage{
parseStratifiers(primary_factor, factors, stratify, facet.x, facet.y)
}
\arguments{
\item{primary_factor}{The name of the primary factor being analyzed}

\item{factors}{List of all factors (with their attributes) of the dataset}

\item{stratify}{Is stratification being performed? If not, this function does
nothing}

\item{facet.x}{(Optional) Name of a factor for horizontal faceting}

\item{facet.y}{(Optional) Name of a factor for vertical faceting}
}
\value{
A nested list of the facets and facet text to be used by MicroVis
    plotting functions for determining faceting and file autonaming
}
\description{
Parse Stratifiers
}
