% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisAlphaDiv.R
\name{phyloDiversity}
\alias{phyloDiversity}
\title{Faith's Phylogenetic Diversity Index}
\usage{
phyloDiversity(dataset = NULL, rooted = F)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset}

\item{rooted}{Calculate rooted or unrooted Faith's phylogenetic diversity index?
Defaults to FALSE}
}
\value{
Single data.frame column of Faith's phylogenetic diversity index values
    for each sample
}
\description{
Faith's Phylogenetic Diversity Index
}
