% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAlphaDiversity.R
\name{plotAlphaDiv}
\alias{plotAlphaDiv}
\title{Plot Alpha Diversity}
\usage{
plotAlphaDiv(
  dataset = NULL,
  method = "Chao1",
  factor = NULL,
  stratify = FALSE,
  facet.x = NULL,
  facet.y = NULL,
  flattenFactors = FALSE,
  param = FALSE,
  showStats = TRUE,
  width = 6,
  height = 4.5,
  add_xaxis = F,
  separateLegend = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{method}{Method for alpha diversity calculation. One of either "chao1",
"shannon", "simpson", "invsimpson", or "pd". Defaults to "chao1"}

\item{factor}{Factor along which to compare alpha diversity between groups.
Defaults to the active factor}

\item{stratify}{Whether or not to stratify. Defaults to FALSE}

\item{facet.x}{(Optional) A factor to stratify by horizontally}

\item{facet.y}{(Optional) A factor to stratify by vertically}

\item{flattenFactors}{Combine multiple factors}

\item{param}{Whether to perform parametrized or nonparametrized univariate
analysis of alpha diversity. Defaults to FALSE (nonparametrized)}

\item{showStats}{Whether or not to show significance labels. Defaults to TRUE}

\item{width}{Width of plot when saved, in inches. Increasing width and height will make text relatively smaller}

\item{height}{Height of plot when saved, in inches. Increasing width and height will make text relatively smaller}

\item{add_xaxis}{Whether to add x-axis labels. Defaults to FALSE}

\item{separateLegend}{Whether to separate the legend from the plot. Defaults
to FALSE}
}
\value{
Boxplot of alpha diversity of the groups
}
\description{
Plot Alpha Diversity
}
