% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBetaDiversity.R
\name{plotBetaDiv}
\alias{plotBetaDiv}
\title{Plot Beta Diversity}
\usage{
plotBetaDiv(
  dataset = NULL,
  method = "bray",
  weighted = F,
  factor = NULL,
  stratify = F,
  spokes = T,
  ellipse = T,
  ci = 0.95,
  showStats = T,
  labelSamples = F,
  separateLegend = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{method}{Method for dissimilarity calculation. One of either "bray",
"euclidean", "jaccard", "unifrac", "manhattan", "canberra", "clark",
"kulczynski", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao", "mahalanobis", "chisq" or "chord"}

\item{weighted}{(Optional) If method is set to "unifrac", whether to perform
weighted or unweighted unifrac. Defaults to FALSE}

\item{factor}{Factor along which to group samples for analysis. Defaults to
the active factor}

\item{stratify}{Another factor to stratify groups}

\item{spokes}{Draw spokes from group centroids to each sample? Defaults to TRUE}

\item{ellipse}{Draw a confidence interval ellipse for each group? Defaults to
TRUE}

\item{ci}{Confidence interval level for ellipses (value between 0 and 1).
Defaults to 0.95}

\item{showStats}{Show statistical results in the plot. Defaults to TRUE}

\item{labelSamples}{Label the samples. Defaults to FALSE}

\item{separateLegend}{Separate the lenged from the figure. Defaults to FALSE}
}
\value{
Scatter plot of the dissimilarity indices of the samples of each group
}
\description{
Plot Beta Diversity
}
