% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClad.R
\name{plotClad}
\alias{plotClad}
\title{Plot Cladogram}
\usage{
plotClad(dataset = NULL, factor = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the activedataset}

\item{factor}{Factor to analyze by. Defaults to the active factor}
}
\value{
Cladogram plot of LEfSe significant taxa
}
\description{
Plot Cladogram
}
