% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions.R
\name{plotFeatureDensity}
\alias{plotFeatureDensity}
\title{Plot distribution curves of abundances of each feature across all samples}
\usage{
plotFeatureDensity(dataset = NULL, byFeature = T, rank = NULL, ftlist = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Default is the active dataset}

\item{byFeature}{Split curves by feature?}

\item{rank}{Rank at which to select features to plot}

\item{ftlist}{Features to plot}
}
\value{
Density plots of each feature across all samples
}
\description{
Plot distribution curves of abundances of each feature across all samples
}
