% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureCorr.R
\name{plotFtCorlines}
\alias{plotFtCorlines}
\title{Plot correlation trendlines between different features}
\usage{
plotFtCorlines(
  dataset1 = NULL,
  dataset2 = NULL,
  method = c("pearson", "kendall", "spearman"),
  factor = NULL,
  features1 = NULL,
  features2 = NULL,
  sigsOnly = T,
  alpha = 0.05,
  r_cutoff = 0.5
)
}
\arguments{
\item{dataset1}{MicroVis dataset. Defaults to the active dataset}

\item{dataset2}{(Optional) MicroVis dataset with features to correlate with
the dataset1}

\item{method}{Correlation method. One of either "pearson", "spearman", or
"kendall". Defaults to "spearman"}

\item{factor}{Factor to group samples by}

\item{features1}{(Optional) Specific features to consider from first dataset.
If only one dataset is being used, then these features will be correlated
with themselves}

\item{features2}{(Optional) Specific features to consider from second dataset,
If no second dataset is provided, then `features1` will be correlated with
`features2` from `dataset1`}

\item{sigsOnly}{Only plot significant correlations (after adjusting for multiple
comparisons)? Defaults to TRUE}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{r_cutoff}{Correlation coefficient threshold. Defaults to 0 (all correlation
values will be accepted when choosing which correlations to plot)}
}
\description{
Plot correlation trendlines between different features
}
