% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFeatureCorr.R
\name{plotFtCormat}
\alias{plotFtCormat}
\title{Plot feature correlation matrix}
\usage{
plotFtCormat(
  dataset1 = NULL,
  dataset2 = NULL,
  method = c("spearman", "pearson", "kendall"),
  factor = NULL,
  rank = NULL,
  sigsOnly = T,
  alpha = 0.05,
  adjustp = T,
  r_cutoff = 0,
  showlowrs = F,
  features1 = NULL,
  features2 = NULL,
  matchFts = F,
  theme = c("classic", "modern"),
  circles = F
)
}
\arguments{
\item{dataset1}{MicroVis dataset. Defaults to the active dataset}

\item{dataset2}{(Optional) MicroVis dataset with features to correlate with
the dataset1}

\item{method}{Correlation method. One of either "pearson", "spearman", or
"kendall". Defaults to "spearman"}

\item{factor}{Factor to group samples by}

\item{rank}{Rank of features to use for calculation}

\item{sigsOnly}{Select only significant correlations? Defaults to TRUE}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{adjustp}{Use the adjusted p-values? Defaults to TRUE}

\item{r_cutoff}{Only select correlations with an absolute R-value above a certain
threshold? This value must be between -1 and 1. Defaults to 0}

\item{showlowrs}{(If r_cutoff is non-zero) Still shade correlations that are
not above the correlation cutoff (r_cutoff)? Defaults to F}

\item{features1}{(Optional) Specific features to consider from first dataset.
If only one dataset is being used, then these features will be correlated
with features of the same dataset}

\item{features2}{(Optional) Specific features to consider from second dataset.}

\item{matchFts}{Show the same features in the correlation matrices for all groups?
This can be helpful when comparing correlations between groups. Defaults
to FALSE}

\item{theme}{Coloring theme. One of "classic" or "modern". Defaults to classic}

\item{circles}{Use circles instead of squares to make the matrix? Defaults to
FALSE}
}
\value{
List of the one or two datasets used for this plot
}
\description{
Plot feature correlation matrix
}
