% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGroupedBars.R
\name{plotGroupedBars}
\alias{plotGroupedBars}
\title{Plot grouped bar plot of significant taxa at each rank of all groups}
\usage{
plotGroupedBars(
  dataset = NULL,
  abun_thresh = 0,
  plotSigs = T,
  ftlist = NULL,
  addfts = NULL,
  factor = NULL,
  stratify = FALSE,
  flattenFactors = FALSE,
  facet.x = FALSE,
  facet.y = FALSE,
  alpha = 0.05,
  param = F,
  byrank = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{abun_thresh}{Abundance cutoff for what features to show. Defaults to 0
(no threshold)}

\item{plotSigs}{Plot significant features only. Defaults to TRUE}

\item{ftlist}{(Optional) List of specific features to plot}

\item{addfts}{(Optional) Additional features to plot in addition to significant
ones}

\item{factor}{Factor to perform statistical analysis by. Defaults to the
active factor}

\item{stratify}{Stratify by another factor? Defaults to FALSE}

\item{flattenFactors}{Whether to combine factors. Defaults to FALSE}

\item{facet.x}{(Optional) Factor to stratify horizontally}

\item{facet.y}{(Optional) Factor to stratify vertically}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{param}{Whether to perform parametrized or nonparametrized univariate
analysis. Defaults to FALSE (nonparametrized)}

\item{byrank}{Facet the plot by rank? Defaults to FALSE}
}
\value{
Grouped bar plot of significant features
}
\description{
Plot grouped bar plot of significant taxa at each rank of all groups
}
