% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap of Data}
\usage{
plotHeatmap(
  dataset = NULL,
  clustNum = 0,
  ftlist = NULL,
  plotSigs = T,
  plotUniques = F,
  param = F,
  aggregated = F,
  labelFeatures = T,
  labelSamples = F,
  factor = NULL,
  stratify = F,
  scaleCol = 1,
  alpha = 0.05,
  width = 6,
  height = 12
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{clustNum}{Number of clusters for rows (features)}

\item{ftlist}{(Optional) List of specific features to plot}

\item{plotSigs}{Plot only significant features? Defaults to TRUE}

\item{plotUniques}{Plot only features that are uniquely expressed in a certain
group? Defaults to FALSE}

\item{param}{Perform parametrized or nonparametrized univariate analysis for
significance testing? Defaults to FALSE}

\item{aggregated}{Aggregate samples in the same group into a single column?
Defaults to FALSE}

\item{labelFeatures}{Label rows with the feature names? Defaults to TRUE}

\item{labelSamples}{Label the columns with the sample names (at the bottom)?
Defaults to FALSE}

\item{factor}{Factor to perform univariate analysis on}

\item{stratify}{Stratify the groups by another factor? Defaults to FALSE}

\item{scaleCol}{Color scaling to improve contrast between certain cells. Value
should be 0 or more. Defaults to 0 (no scaling)}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{width}{Width of saved heatmap figure in inches. Defaults to 6 in}

\item{height}{Height of saved heatmap figure in inches. Defaults to 8 in}
}
\value{
Heatmap of relative abundances of features in each group
}
\description{
Plot Heatmap of Data
}
