% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatTree.R
\name{plotHeatTree}
\alias{plotHeatTree}
\title{Plot Heat Tree}
\usage{
plotHeatTree(
  dataset = NULL,
  test = "univar",
  ftlist = NULL,
  showAll = F,
  color_limit = 2,
  alpha = 0.05,
  sigsOnly = T,
  initial_layout = "fruchterman-reingold",
  layout = "reingold-tilford"
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{test}{Type of statistical test to use. One of either "univar" or "deseq".
Defaults to univariate}

\item{ftlist}{(Optional) List of specific features to show}

\item{showAll}{Show all features? Defaults to FALSE}

\item{color_limit}{Any log2FC values with absolute value above this will be
treated as if they are at this value for coloring}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{initial_layout}{Initial layout for the tree}

\item{layout}{Type of tree layout}
}
\value{
MicroVis dataset
}
\description{
Plot Heat Tree
}
