% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLefse.R
\name{plotLEFSE}
\alias{plotLEFSE}
\title{Plot LEfSe analysis (not cladogram plot)}
\usage{
plotLEFSE(
  dataset = NULL,
  factor = NULL,
  alpha = 0.05,
  lda_cutoff = 2,
  top = 20,
  byrank = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{Factor to analyze data by. Defaults to the active factor}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{lda_cutoff}{Effect size cutoff for features to plot. Defaults to 2}

\item{top}{Top number of features by effect size to plot. Defaults to 20}

\item{byrank}{Facet plot by rank? Defaults to FALSE}
}
\value{
Plot of effect sizes of significant features from LEfSe analysis
}
\description{
Plot LEfSe analysis (not cladogram plot)
}
