% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot Feature Network}
\usage{
plotNetwork(
  dataset = NULL,
  factor = NULL,
  rank = NULL,
  fts = NULL,
  method = c("se.mb", "se.glasso"),
  format = c("ggplot", "igraph"),
  fill = NULL,
  outline = NULL,
  labelfts = NULL,
  labelAll = F,
  deg_cutoff = 0,
  r_cutoff = 0,
  top_r_prop = 100,
  layout = c("fr", "circle", "sphere", "dh", "nicely")
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{(Optional) Factor to split data along groups}

\item{rank}{Rank at which to build network}

\item{fts}{List of specific features to analyze and plot}

\item{method}{Co-occurence analysis method}

\item{format}{Use ggplot or igraph to plot the network?}

\item{fill}{How node colors are chosen}

\item{outline}{(Optional) How node outlines are chosen, if at all. Defaults to
no outline}

\item{labelfts}{Which nodes are labeled, if any. Defaults to none}

\item{labelAll}{Label all features in the network? Defaults to FALSE}

\item{deg_cutoff}{Only display nodes with at least this many connections/edges.
Defaults to 0}

\item{r_cutoff}{Only display connections/edges with an absolute value of this or more}

\item{top_r_prop}{Only display the top percentile of connections by absolute value}

\item{layout}{Layout for nodes}
}
\value{
Network plot
}
\description{
Plot Feature Network
}
