% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrelations.R
\name{plotPairedCor}
\alias{plotPairedCor}
\title{Correlate abundance of features between paired samples}
\usage{
plotPairedCor(dataset = NULL, ids, compare, fts = NULL, invert = F)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{ids}{Column(s) that uniquely identify each pair of samples}

\item{compare}{Column with the 2 groups to correlate between. Each pair of
samples has a sample in each of the two groups in this column. If there
are more than 2 groups in the column, user will be asked to select 2 to
compare}

\item{fts}{Vector of features to correlate between the two sets}
}
\value{
Scatter plot with trendline
}
\description{
Correlate abundance of features between paired samples
}
