% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRareCurves.R
\name{plotRareCurves}
\alias{plotRareCurves}
\title{Plot rarefaction curves}
\usage{
plotRareCurves(
  dataset = NULL,
  step_size = 100,
  raw = F,
  panels = TRUE,
  labelSamples = FALSE,
  getPlot = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{step_size}{Step size for drawing the curves}

\item{raw}{Perform on unfiltered data?}

\item{panels}{Split groups into separate panels?}

\item{labelSamples}{Label each curve with the sample it corresponds to?}

\item{getPlot}{Return the plot object?}
}
\value{
Either the MicroVis dataset or the rarefaction curve plot object
}
\description{
Plot rarefaction curves
}
