% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRFBoruta.R
\name{plotRF}
\alias{plotRF}
\title{Plot random forest important features}
\usage{
plotRF(
  dataset = NULL,
  factor = NULL,
  hideShadow = T,
  confirmedOnly = F,
  top = 15,
  max_runs = 100,
  roughfix = F,
  alpha = 0.01
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{Factor to analyze by}

\item{hideShadow}{Hide Boruta shadow features? Defaults to TRUE}

\item{confirmedOnly}{Only plot confirmed features? Defaults to FALSE}

\item{top}{How many top features to show. Defaults to 15}

\item{max_runs}{Max Boruta runs. Defaults to 100}

\item{roughfix}{Roughfix indeterminate Boruta features? Defaults to FALSE}

\item{alpha}{Significance threshold. Defaults to 0,01}
}
\value{
Plot of important features determined by the Boruta algorithm on
    a random forest model
}
\description{
Plot random forest important features
}
