% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDistributions.R
\name{plotSampleDensity}
\alias{plotSampleDensity}
\title{Plot distribution curves of feature abundances in each sample}
\usage{
plotSampleDensity(dataset = NULL, bySample = T)
}
\arguments{
\item{dataset}{MicroVis dataset. Default is the active dataset}

\item{bySample}{Split curves by sample?}
}
\value{
Density plots of feature abundances of each sample
}
\description{
Plot distribution curves of feature abundances in each sample
}
