% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackedBars.R
\name{plotStackedBars}
\alias{plotStackedBars}
\title{Plot stacked abundace barplots}
\usage{
plotStackedBars(
  dataset = NULL,
  proportional = T,
  factor = NULL,
  stratify = F,
  bySample = F,
  unfiltered = T,
  rank = NA,
  top = 15,
  top_by = "max",
  ftlist = c(),
  plotSigs = F,
  alpha = 0.05
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{proportional}{Whether to plot relative/proportional abundance. Defaults
to TRUE}

\item{factor}{Factor to group samples by}

\item{stratify}{Whether to stratify the groups. Defaults to FALSE}

\item{bySample}{Do not aggregate samples by groups and instead show a bar for
each sample. Defaults to FALSE}

\item{unfiltered}{Whether to use data before or after filtering out features.
Defaults to TRUE (uses data before filtering)}

\item{rank}{Rank at which to use features}

\item{top}{Number of features to show based on greatest relative abundance.
Defaults to top 15 features by relative abundance}

\item{top_by}{Select the top features either by the max, min, or mean relative
abundance in the groups, or the summed relative abundance across all groups.
Defaults to max (in other words, for each feature the maximum of its
relative abundance values across the groups will be used to determine
whether it is a top feature).}

\item{ftlist}{(Optional) List of features to consider. Defaults to all features
at the rank}

\item{plotSigs}{Plot only the significant features? Defaults to FALSE}

\item{alpha}{Significance threshold for selecting significant features. Defaults
to 0.05}
}
\value{
Stacked bar plot of feature abundances of selected features with the
    other features lumped into an "Other" category
}
\description{
Plot stacked abundace barplots
}
