% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUnivar.R
\name{plotUnivar}
\alias{plotUnivar}
\title{Plot box-plots of univariate analysis}
\usage{
plotUnivar(
  dataset = NULL,
  jitter = T,
  dotplot = F,
  violin = F,
  showStats = T,
  raw = F,
  factor = NULL,
  stratify = F,
  facet.x = NULL,
  facet.y = NULL,
  flattenFactors = F,
  rank = NULL,
  ftlist = NULL,
  unique_groups = NULL,
  plotAll = F,
  alpha = 0.05,
  param = F,
  scalePlot = F,
  add_xaxis = F,
  separateLegend = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{jitter}{Show "jitter" points for each sample? Defaults to TRUE}

\item{dotplot}{Create dotplot? Defaults to FALSE}

\item{violin}{Create violinplot? Defaults to FALSE}

\item{showStats}{Show bars w/asterisks for significance labels? Defaults to
TRUE}

\item{raw}{Show raw abundance values? Defaults to FALSE. If TRUE, showStats
is forced to FALSE}

\item{factor}{Factor along which to analyze data}

\item{stratify}{Stratify the data along a second factor? Defaults to FALSE}

\item{facet.x}{(Optional) Name of a factor to stratify horizontally}

\item{facet.y}{(Optional) Name of a factor to stratify vertically}

\item{flattenFactors}{(Not ready) Combine two factors into one}

\item{rank}{Rank at which to select features}

\item{ftlist}{List of specific features to plot. Defaults to all features at
the given rank}

\item{unique_groups}{(Optional) Only plot features that are uniquely over/under-
expressed in these groups}

\item{plotAll}{Plot all features regardless of significance? Defaults to FALSE}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{param}{Perform parametrized analysis? Defaults to FALSE (non-parametrized
analysis by default)}

\item{scalePlot}{Scale plot by log2? Defaults to FALSE}

\item{add_xaxis}{Add group labels to the x-axis? Defaults to FALSE}

\item{separateLegend}{Make one separate figure of just the legend instead of
including it in all the boxplots? Defaults to FALSE}
}
\value{
MicroVis dataset
}
\description{
Plot box-plots of univariate analysis
}
