% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsState.R
\name{poolGroups}
\alias{poolGroups}
\title{Pool groups of a factor into fewer, larger groups}
\usage{
poolGroups(dataset = NULL, factor = NULL, include_prefix = T)
}
\arguments{
\item{dataset}{Microvis dataset. Default is the active dataset}

\item{factor}{Groups of this factor will be grouped}

\item{include_prefix}{Include the factor name as the prefix for each group?}
}
\value{
Dataset with a new factor of pooled groups
}
\description{
Pool groups of a factor into fewer, larger groups
}
