% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settingSwitches.R
\name{posthoctoggle}
\alias{posthoctoggle}
\title{Switch between Tukey/Games-Howell or T-test/Wilcox for post-hoc analysis}
\usage{
posthoctoggle()
}
\value{
Does not return anything
}
\description{
Switch between Tukey/Games-Howell or T-test/Wilcox for post-hoc analysis
}
