% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmvdata.R
\name{print.mvdata}
\alias{print.mvdata}
\title{Print function for mvdata (MicroVis dataset) objects}
\usage{
\method{print}{mvdata}(x, ...)
}
\arguments{
\item{x}{MicroVis dataset (mvdata object)}

\item{...}{Arguments passed to print() function}
}
\value{
Print-friendly printout of dataset details
}
\description{
Print function for mvdata (MicroVis dataset) objects
}
