% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printmvdata.R
\name{print.mvmerged}
\alias{print.mvmerged}
\title{Print function for mvmerged (merged MicroVis dataset) objects}
\usage{
\method{print}{mvmerged}(x, ...)
}
\arguments{
\item{x}{Merged MicroVis dataset (mvmerged object)}

\item{...}{Arguments passed to print() function}
}
\value{
Print-friendly printout of dataset details
}
\description{
Print function for mvmerged (merged MicroVis dataset) objects
}
