% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMVDataset.R
\name{processDataset}
\alias{processDataset}
\title{Process Dataset for Analysis}
\usage{
processDataset(dataset, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{If TRUE, will not set the processed dataset as the active dataset.}

\item{silent}{If TRUE, none of the processing steps will print out.}
}
\value{
Processed MicroVis dataset (mvdata object)
}
\description{
Called by a number of functions to process a dataset based on
    set parameters.
}
