% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarefyData.R
\name{rarefySamples}
\alias{rarefySamples}
\title{Data Rarefaction}
\usage{
rarefySamples(dataset = NULL, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{This is not needed}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
MicroVis dataset (mvdata object) with updated rarefaction parameters
}
\description{
Data Rarefaction
}
