% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDataset.R
\name{readdepth}
\alias{readdepth}
\title{Get read depth summary statistics for each group or for all samples}
\usage{
readdepth(dataset = NULL, factor = NULL, individual = F, all = F)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}

\item{factor}{(Optional) Factor along which samples will be grouped}

\item{individual}{(Optional) Whether to return read depths for each individual
sample. Defaults to FALSE}

\item{all}{(Optional) Whether to calculate read depth statistics for all the
samples as one group}
}
\value{
Dataframe of read depth statistics including min, max, median, and mean
}
\description{
Get read depth summary statistics for each group or for all samples
}
