% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFiltering.R
\name{removeGrps}
\alias{removeGrps}
\title{Remove Groups}
\usage{
removeGrps(dataset = NULL, factor_name = NULL, grps = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{factor_name}{(Optional) Names of the factors for which to specify groups}

\item{grps}{(Optional) Names of the groups to remove}
}
\value{
MicroVis dataset (mvdata object) with updated "subset" lists for each
    of its factors, and an updated list of samples for runSampleFilter() to
    remove based on the groups chosen for analysis
}
\description{
Remove Groups
}
