% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFiltering.R
\name{removeLowQuality}
\alias{removeLowQuality}
\title{Remove Low Quality Samples}
\usage{
removeLowQuality(dataset = NULL, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runFeatureFilter(), and runNormalizer(). If TRUE, it tells these functions
to NOT print out any of their processing messages.}
}
\value{
MicroVis dataset (mvdata object) with an updated list of low-quality
    features for runSampleFilter() to filter out
}
\description{
Remove Low Quality Samples
}
