% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsState.R
\name{removePool}
\alias{removePool}
\title{Remove a factor of pooled groups}
\usage{
removePool(dataset = NULL)
}
\arguments{
\item{dataset}{Dataset to remove pooling factor from}
}
\value{
Returns a dataset without the pooling factor that was selected for
    for removal
}
\description{
Remove a factor of pooled groups
}
