% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultSaving.R
\name{resetResDir}
\alias{resetResDir}
\title{Reset the Results Directory}
\usage{
resetResDir(save_directory)
}
\arguments{
\item{save_directory}{Current results directory path}
}
\value{
New results directory path.
}
\description{
Reset the Results Directory
}
