% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRF-Boruta.R
\name{rfboruta}
\alias{rfboruta}
\title{Perform random forest Boruta analysis}
\usage{
rfboruta(
  dataset = NULL,
  factor = NULL,
  max_runs = 100,
  roughfix = F,
  alpha = 0.01
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{Factor to analyze by. Defaults to the active factor}

\item{max_runs}{Maximum number of runs for Boruta algorithm. Default is 100}

\item{roughfix}{Perform roughfix of indeterminate features? Defaults to FALSE}

\item{alpha}{Significance threshold. Defaults to 0.01}
}
\value{
List of 3 tables with the class error, errors, and the Boruta results
}
\description{
Perform random forest Boruta analysis
}
