% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{runFeatureFilter}
\alias{runFeatureFilter}
\title{Feature Filter}
\usage{
runFeatureFilter(dataset = NULL, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{This parameter has no use in this function and can be removed}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
MicroVis dataset (mvdata object) with features filtered out based on
    parameters set by the filter___ feature filtering functions
}
\description{
This function is one of the processing functions called by
    processDataset() and identifies and removes features based on parameters
    set by the filter___ feature filtering functions
}
