% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeFeatures.R
\name{runFeatureRemover}
\alias{runFeatureRemover}
\title{Feature Remover}
\usage{
runFeatureRemover(dataset = NULL, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{This parameter has no use in this function and can be removed}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
MicroVis dataset (mvdata object) with features filtered out based on
    parameters set by the filter___ feature filtering functions
}
\description{
This function removes features prior to any processing and is
    one of the processing functions called by processDataset(). It should be
    used WITH CAUTION as it can drastically modify a dataset, making it
    completely unrepresentative if used incorrectly.
}
