% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{runNormalization}
\alias{runNormalization}
\title{Data Normalization}
\usage{
runNormalization(dataset = NULL, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{This parameter has no use in this function and can be removed}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
MicroVis dataset (mvdata object) with normalized data
}
\description{
Normalizes unranked data based on parameters set by the
    scaleSamples, scaleFeatures, and transData. This function is one of the
    processing functions called by processDataset().
}
