% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFiltering.R
\name{runSampleFilter}
\alias{runSampleFilter}
\title{Sample Filter}
\usage{
runSampleFilter(dataset = NULL, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{This parameter has no use in this function and can be removed}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runFeatureFilter(), and runNormalizer(). If TRUE, it tells these functions
to NOT print out any of their processing messages.}
}
\value{
MicroVis dataset (mvdata object) with the following samples removed:
    1) Low quality samples identified by removeLowQuality()
    2) Samples in groups excluded by chooseGrps()/addGrps()/removeGrps()
    3) Specific samples removed by the user using removeSamples()
}
\description{
This function is one of the processing functions called by
    processDataset() removes samples from low quality samples identified by
    removeLowQuality(), groups excluded by chooseGrps()/addGrps()/removeGrps()
    , and samples specifically excluded by removeSamples()
}
