% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFrequencies.R
\name{sampleFreq}
\alias{sampleFreq}
\title{Frequency table of samples}
\usage{
sampleFreq(
  dataset = NULL,
  characteristics,
  unique = NULL,
  subset = NULL,
  raw = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{characteristics}{A vector of 2 column names in the metadata with which to create the frequency table.
First variable will be rows and second will be columsn.}

\item{unique}{(Optional) Column to use to select unique samples}

\item{subset}{(Optional) A named list where names are additional column names, and their values are values to subset the samples by.}

\item{raw}{Whether to create frequency table from unprocessed data so that any samples excluded due to low quality or other sample filtering are included.
Defaults to False.}
}
\value{
Returns a frequency table of two @characteristics
}
\description{
Frequency table of samples
}
