% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesWith.R
\name{samplesWith}
\alias{samplesWith}
\title{Filter a datatable to only samples with certain metadata characteristics}
\usage{
samplesWith(data, id_cols, with_cols)
}
\arguments{
\item{data}{Data table with metadata}

\item{id_cols}{Vector of names of columns that uniquely identify subjects}

\item{with_cols}{Vector of names of columns that each subject must have all of}
}
\value{
Data table with just the samples identifiedy by "id_cols" that have
    all the values in with_cols
}
\description{
Filter a datatable to only samples with certain metadata characteristics
}
