% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settingSwitches.R
\name{savetoggle}
\alias{savetoggle}
\title{Switch Saving Mode}
\usage{
savetoggle()
}
\value{
Does not return anything
}
\description{
Cycles between three options for saving results of plotting:
    - Offer to save after each plotting function is run
    - Automatically save without asking
    - Do not save anything and don't ask the user either
}
