% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{scaleFeatures}
\alias{scaleFeatures}
\title{Scale Features}
\usage{
scaleFeatures(
  dataset = NULL,
  temp = F,
  scaling = c("sum", "median", "range", "centered", "standardized", "relative", "none"),
  sum_scale = 1e+06,
  silent = F
)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{If set to TRUE, it tells processDataset() to NOT update the active
dataset.}

\item{scaling}{Method for scaling features. Choices are "sum", "none", "relative",
"median", "range", "centered", and "standardized". Default is "sum".}

\item{sum_scale}{Number of reads to scale each feature to. Default is 1 million.}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
Microvis dataset (mvdata object) with features scaled.
}
\description{
Scale Features
}
