% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{scaleSamples}
\alias{scaleSamples}
\title{Scale Samples}
\usage{
scaleSamples(
  dataset = NULL,
  temp = F,
  scaling = "sum",
  sum_scale = 1e+06,
  silent = F
)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{If set to TRUE, it tells processDataset() to NOT update the active
dataset.}

\item{scaling}{Method for scaling samples. Choices are "sum", "none", "relative",
"median", "range", "centered", and "standardized". Default is "sum".}

\item{sum_scale}{Number of reads to scale each sample to. Default is 1 million.}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
Microvis dataset (mvdata object) with samples scaled.
}
\description{
Scale Samples
}
