% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{selectFeatures}
\alias{selectFeatures}
\title{Select Specific Features}
\usage{
selectFeatures(dataset = NULL, features, ranks = NULL, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{features}{Vector of features (at any rank) to select}

\item{ranks}{(Optional) Specify a rank from which to select features. Searches
through all ranks by default}

\item{temp}{This parameter has no use in this function and can be removed}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
Dataset with list of selected features that is passed to runFeatureFilter
}
\description{
Select Specific Features
}
