% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setFactorVariable.R
\name{setFVar}
\alias{setFVar}
\title{Set Factor Variable}
\usage{
setFVar(dataset, factor_name = NULL)
}
\arguments{
\item{dataset}{Microvis dataset (mvdata object)}

\item{factor_name}{(Optional) Name of the desired factor}
}
\value{
Factor variable (a list object) with its name, groups, subsetted groups,
    and print-friendly name (spaces instead of underscores)
}
\description{
This function just assigns a factor in the dataset to a variable
    to be used within a function. If no factor name is specified in the function
    call, it will choose the dataset's active factor.
}
