% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisFeatureCorr.R
\name{sliceCormats}
\alias{sliceCormats}
\title{Select only significant and/or high-correlation values from correlation matrices}
\usage{
sliceCormats(
  cormats,
  sigsOnly = T,
  alpha = 0.05,
  adjustp = T,
  r_cutoff = 0,
  matchFts = F
)
}
\arguments{
\item{cormats}{List of correlation matrices (output of ftcor)}

\item{sigsOnly}{Select only significant correlations? Defaults to TRUE}

\item{alpha}{Significance threshold. Defaults to 0.05}

\item{adjustp}{Use the adjusted p-values? Defaults to TRUE}

\item{r_cutoff}{Only select correlations with an absolute R-value above a certain
threshold? This value must be between -1 and 1. Defaults to 0}

\item{matchFts}{Show the same features in the correlation matrices for all groups?
This can be helpful when comparing correlations between groups. Defaults
to FALSE}
}
\value{
List of correlation matrices with only the desired correlations remaining
}
\description{
Select only significant and/or high-correlation values from correlation matrices
}
