% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeTaxa.R
\name{summarizeTaxa}
\alias{summarizeTaxa}
\title{Get mean and SE of features}
\usage{
summarizeTaxa(
  dataset = NULL,
  factor = NULL,
  ranks = NULL,
  allRanks = F,
  meansOnly = F,
  pivoted = T
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset}

\item{factor}{Factor along which to conduct beta-diversity analysis}

\item{ranks}{Ranks at which to summarize features}

\item{allRanks}{If set to TRUE, will summarize features at all ranks. Defaults
to FALSE}

\item{meansOnly}{Only return the means and not SE}

\item{pivoted}{If set to TRUE, the returned table will have each group of the
factor in its own column}
}
\value{
Table with the means +/- SE for features in each group of the dataset
}
\description{
Get mean and SE of features
}
