% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ASVtoTaxa.R
\name{TaxatoASV}
\alias{TaxatoASV}
\title{Translate Taxa Names to ASV Numbers}
\usage{
TaxatoASV(taxa_data, taxalist, taxa_rank = NULL)
}
\arguments{
\item{taxa_data}{Taxonomy data (list object) from a dataset containing
asv-taxonomy reference table.}

\item{taxalist}{List of taxa names to look up in taxa_data.}

\item{taxa_rank}{(Optional) The desired rank at which to look up taxa names.
Will default to the active rank if not specified.}
}
\value{
List of ASVs corresponding to the taxa in the same order.
}
\description{
Translate Taxa Names to ASV Numbers
}
