% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisUnivar.R
\name{univar}
\alias{univar}
\title{Perform Univariate Analysis}
\usage{
univar(
  dataset = NULL,
  data = NULL,
  factor = NULL,
  stratifiers = NULL,
  rank = NULL,
  features = NULL,
  pairwise_comparisons = NULL,
  param = F,
  dataset_name = NULL
)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset. Defaults to the active dataset.
Must either provide a "dataset" or "data" argument}

\item{data}{(Optional) Data table containing metadata and count values with samples as
rows and metadata factors or features as columns}

\item{factor}{Factor along which to form groups from the samples and perform
the statistical analysis}

\item{stratifiers}{(Optional) One or two factors to stratify the groups by}

\item{rank}{Rank at which to perform the dataset. Defaults to the active rank}

\item{features}{List of features to analyze. Defaults to all}

\item{pairwise_comparisons}{For 3 or more groups, list of vectors of pairwise
comparisons to conduct. Defaults to all possible pairwise comparisons}

\item{param}{Run parametrized or nonparametrized univariate analysis? Default
is FALSE}

\item{dataset_name}{(Not recommended) Name of the dataset to save statistics
to. This should not need to be used by users since the function can
determine the name of the dataset directly passed to it, but not when
it is called within another function.}
}
\value{
MicroVis dataset containing the statistics results in its "stats"
    attribute
}
\description{
Perform Univariate Analysis
}
