% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{unselectFeatures}
\alias{unselectFeatures}
\title{Undo Feature Selection}
\usage{
unselectFeatures(dataset = NULL, temp = F, silent = F)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{temp}{This parameter has no use in this function and can be removed}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
MicroVis dataset with the selected feature list cleared
}
\description{
Undo Feature Selection
}
