% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changeFactor.R
\name{validFactor}
\alias{validFactor}
\title{Factor Validation}
\usage{
validFactor(dataset)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}
}
\value{
Dataset with a new active factor.
}
\description{
Function for checking factors to make sure they are valid for
    analysis. If the currently active factor only has 1 group, a new active
    factor will be chosen.
}
