% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDataset.R
\name{viewstats}
\alias{viewstats}
\title{View Statistics of All Features}
\usage{
viewstats(dataset = NULL, type = NULL, factor = NULL, rank = NULL)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}

\item{type}{Type of statistical test to view results for. Defaults to the first
available of univariate, deseq, and lefse in that order}

\item{factor}{(Optional) Factor to count samples by groups. Default is the
active factor of the dataset.}

\item{rank}{(Optional) Rank to get abundance values of. Default is active rank.}
}
\value{
View of the statistics of the dataset
}
\description{
View Statistics of All Features
}
