% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspectDataset.R
\name{viewtable}
\alias{viewtable}
\title{View Abundance Table}
\usage{
viewtable(dataset = NULL)
}
\arguments{
\item{dataset}{(Optional) MicroVis dataset (mvdata object). If not specified,
defaults to active dataset.}
}
\value{
Does not return anything.
}
\description{
View Abundance Table
}
