% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{zeroReplace}
\alias{zeroReplace}
\title{Replace Zeros with Non-Zero Value}
\usage{
zeroReplace(x)
}
\arguments{
\item{x}{Numerical matrix}
}
\value{
Matrix with no zeroes
}
\description{
Replace Zeros with Non-Zero Value
}
