% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settingSwitches.R
\name{autonametoggle}
\alias{autonametoggle}
\title{Automatically Name Results}
\usage{
autonametoggle(setting = c("groups", "dataset", "manual"))
}
\arguments{
\item{settings}{(Optional) Whether to set naming of files to auto-naming by
groups in analysis, auto-naming by the name of the dataset, or manual
naming by the user. By default, this function cycles through these 3 settings}
}
\value{
Does not return anything
}
\description{
Cycles between naming results automatically or prompting user
    for a name
}
