% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{getFtStats}
\alias{getFtStats}
\title{Get Feature Statistics for Filtering}
\usage{
getFtStats(dataset = NULL, rank = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{rank}{Rank at which to get feature summary stats}
}
\value{
List containing original abundance table, feature names table, and
    a dataframe of feature statistics used for filtering
}
\description{
Get Feature Statistics for Filtering
}
