% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureFiltering.R
\name{getFtStats}
\alias{getFtStats}
\title{Get Feature Statistics for Filtering}
\usage{
getFtStats(dataset = NULL, rank = NULL, raw = T, justStats = T)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{rank}{Rank at which to get feature summary stats}

\item{raw}{Use raw counts to calculate feature statistics. Defaults to TRUE}

\item{justStats}{Whether to just output the a table of feature statistics
instead of all data processing information}
}
\value{
List containing original abundance table, feature names table, and
    a dataframe of feature statistics used for filtering
}
\description{
Get Feature Statistics for Filtering
}
