% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplesWith.R
\name{getSamples}
\alias{getSamples}
\title{Get samples with certain metadata characteristics from a dataset}
\usage{
getSamples(dataset = NULL, id_cols = NULL, complete = NULL, filter = NULL)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{id_cols}{Vector of names of metadata columns that uniquely identify subjects}

\item{complete}{Get samples with all instances (complete cases) of this variable
in the metadata}

\item{filter}{Vector of names of metadata columns that each subject must have all of}
}
\value{
Data table with just the samples identifiedy by "id_cols" that have
    all the values in filter
}
\description{
Get samples with certain metadata characteristics from a dataset
}
