% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvsave}
\alias{mvsave}
\title{Save Dataset in its Current State}
\usage{
mvsave(dataset_name = NULL)
}
\arguments{
\item{dataset_name}{(Optional) MicroVis dataset (mvdata object). Defaults to
active dataset}
}
\value{
MicroVis dataset (mvdata object)
}
\description{
Save Dataset in its Current State
}
