% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mvsave}
\alias{mvsave}
\title{Save Dataset in its Current State}
\usage{
mvsave(name = NULL)
}
\arguments{
\item{name}{New name to save the currently active dataset
to. If none is specified, simply returns the dataset}
}
\value{
MicroVis dataset (mvdata object)
}
\description{
Save Dataset in its Current State
}
