% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNetwork.R
\name{plotNetwork}
\alias{plotNetwork}
\title{Plot Feature Network}
\usage{
plotNetwork(
  dataset = NULL,
  factor = NULL,
  rank = NULL,
  fts = NULL,
  method = c("spieceasi", "sparcc"),
  fill = NULL,
  outline = NULL,
  labelFts = NULL,
  labelAll = F,
  layout = c("fr", "circle")
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{factor}{(Optional) Factor to split data along groups}

\item{rank}{Rank at which to build network}

\item{fts}{List of specific features to analyze and plot}

\item{method}{Co-occurence analysis method}

\item{fill}{How node colors are chosen}

\item{outline}{(Optional) How node outlines are chosen, if at all. Defaults to
no outline}

\item{layout}{Layout for nodes}

\item{labelfts}{Which nodes are labeled, if any. Defaults to none}
}
\value{
Network plot
}
\description{
Plot Feature Network
}
