% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStackedBars.R
\name{plotStackedBars}
\alias{plotStackedBars}
\title{Plot stacked abundace barplots}
\usage{
plotStackedBars(
  dataset = NULL,
  relative = T,
  factor = NULL,
  stratify = F,
  bySample = F,
  allSamples = F,
  unfiltered = T,
  label_unknown = T,
  rank = NA,
  top = 15,
  top_by = "max",
  ftlist = c(),
  plotSigs = F,
  alpha = 0.05,
  separateLegend = F
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to the active dataset}

\item{relative}{Whether to plot relative abundance. Defaults to TRUE}

\item{factor}{Factor to group samples by}

\item{stratify}{Whether to stratify the groups. Defaults to FALSE}

\item{bySample}{Whether to show a bar for each sample instead of aggregating
samples by groups. Defaults to FALSE}

\item{allSamples}{Whether to plot a single bar for all the samples in the
dataset. Defaults to FALSE}

\item{unfiltered}{Whether to use data before or after filtering out features.
Defaults to TRUE (uses data before filtering)}

\item{label_unknown}{Whether to group unknown taxa separate. Defaults to TRUE}

\item{rank}{Rank at which to use features}

\item{top}{Number of features to show based on greatest relative abundance.
Defaults to top 15 features by relative abundance}

\item{top_by}{Select the top features either by the max, min, or mean relative
abundance in the groups, or the summed relative abundance across all groups.
Defaults to mean (in other words, for each feature its mean
relative abundance values across the samples will be used to determine
whether it is a top feature).}

\item{ftlist}{(Optional) List of features to consider. Defaults to all features
at the rank}

\item{plotSigs}{Plot only the significant features? Defaults to FALSE}

\item{alpha}{Significance threshold for selecting significant features. Defaults
to 0.05}

\item{separateLegend}{Whether to separate the legend from the plot. Defaults
to FALSE}
}
\value{
Stacked bar plot of feature abundances of selected features with the
    other features lumped into an "Other" category
}
\description{
Plot stacked abundace barplots
}
