% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisPermanova.R
\name{pnova}
\alias{pnova}
\title{PERMANOVA Analysis}
\usage{
pnova(
  dataset = NULL,
  dist = "bray",
  weighted = F,
  allFactors = T,
  factors = NULL,
  alpha = 0.05
)
}
\arguments{
\item{dataset}{MicroVis dataset. Defaults to active dataset}

\item{dist}{Method for dissimilarity calculation. One of either "bray",
"euclidean", "jaccard", "unifrac", "manhattan", "canberra", "clark",
"kulczynski", "gower", "altGower", "morisita", "horn", "mountford", "raup",
"binomial", "chao", "cao", "mahalanobis", "chisq" or "chord"}

\item{weighted}{(Optional) If method is set to "unifrac", whether to perform
weighted or unweighted unifrac. Defaults to FALSE}

\item{allFactors}{Perform PERMANOVA analysis on all factors? Defaults to TRUE
if no factors are specified (below)}

\item{factors}{(Optional) Specify factors to perform PERMANOVA on. If none are
specified, then all factors will be included}

\item{alpha}{Significance threshold for PERMANOVA results before performing
PERMDISP analysis. Defaults to 0.05}
}
\value{
PERMANOVA and PERMDISP results
}
\description{
PERMANOVA Analysis
}
