% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleFiltering.R
\name{removeSamples}
\alias{removeSamples}
\title{Remove Specific Samples}
\usage{
removeSamples(dataset = NULL, samples)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{samples}{Vector of names of the samples to be removed}
}
\value{
MicroVis dataset (mvdata object) with an updated "ignored" lists of
    samples for runSampleFilter() to remove
}
\description{
Remove Specific Samples
}
