% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultSaving.R
\name{saveResults}
\alias{saveResults}
\title{Save figure and statistics results}
\usage{
saveResults(
  save_directory,
  foldername = NULL,
  filename = NULL,
  factors = NULL,
  analysis_name = NULL,
  active_factor = NULL,
  facets = NULL,
  figure = NULL,
  width = 18,
  height = 12,
  stat_results = NULL,
  other_results = NULL,
  suffix = "",
  forcesave = F,
  verbose = T
)
}
\arguments{
\item{save_directory}{Path to directory to save results to}

\item{foldername}{(Optional) Name of the directory within save_directory}

\item{filename}{(Optional) Name of file to save figure and statistics under}

\item{factors}{(Optional) Factors of the dataset}

\item{analysis_name}{(Optional) Name for this analysis}

\item{active_factor}{(Optional) The active factor of this analysis}

\item{facets}{(Optional) Facets if there were any facets in this analysis}

\item{figure}{(Optional) The plot object to be saved}

\item{width}{(Optional) Width of the image to be saved in inches}

\item{height}{(Optional) Height of the image to be saved in inches}

\item{stat_results}{(Optional) Table of statistics results to be saved}

\item{other_results}{(Optional) Table of other results to be saved}

\item{suffix}{(Optional) Any suffix to be added to the end of the file name}

\item{forcesave}{If set to TRUE, the function acts as if "autosave" is TRUE}

\item{verbose}{If set to TRUE, will print out the location of file saving}
}
\value{
save_directory
}
\description{
Save figure and statistics results
}
