% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{transData}
\alias{transData}
\title{Data Transformation}
\usage{
transData(
  dataset = NULL,
  temp = F,
  trans_method = c("clr", "glog", "pseudolog", "log", "gmpr", "none"),
  log_base = 10,
  impute_method = c("GBM", "SQ", "BL", "CZM"),
  silent = F
)
}
\arguments{
\item{dataset}{MicroVis dataset (mvdata object)}

\item{temp}{If set to TRUE, it tells processDataset() to NOT update the active
dataset.}

\item{trans_method}{Method for transforming data. Choices are "glog"
(generalized log), "none", "pseudolog", and "log".}

\item{log_base}{Base of log transformation.}

\item{silent}{Argument that is ultimately passed onto runSampleFilter(),
runNormalization(), and runFeatureFilter(), telling them not to output
any messages.}
}
\value{
Microvis dataset (mvdata object) with samples scaled.
}
\description{
Data Transformation
}
