% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settingSwitches.R
\name{zerostoggle}
\alias{zerostoggle}
\title{Switch Method for Replacing Zeros}
\usage{
zerostoggle(method = c("replace", "impute"), div = NULL)
}
\arguments{
\item{method}{(Optional) The method of zero-replacement to switch to}

\item{div}{(Optional) The divisor by which to divide the lowest count value by
to determine the maximum value to replace zeros with for the "replace" method}
}
\value{
Does not return anything
}
\description{
Cycles between two options for dealing with zero counts:
    - replace: Replacing zero counts with a value between 1/100 and 1/10 of the minimum count value
    - impute: Use zCompositions package to impute a distribution of values to replace zeros
}
